refactor(lambda functions): reorganize functions for better maintainability
