+++ cloudformation/gracychat.yaml
@@ -22,7 +22,7 @@ Parameters:
     Description: S3 key for the Lambda layer ZIP file.
 
 Resources:
-  # DynamoDB Table for Chatbot Logs
+  # DynamoDB Table for Interaction Logs
   ChatbotLogsTable:
     Type: AWS::DynamoDB::Table
     Properties:
@@ -88,7 +88,7 @@ Resources:
     Type: AWS::Lambda::Function
     Properties:
       FunctionName: GracyChatFunction
-      Handler: lambda_function.lambda_handler
+      Handler: functions/lambda_function.lambda_handler
       Role: !GetAtt GracyChatLambdaRole.Arn
       Runtime: python3.13
       Timeout: 30
@@ -124,7 +124,7 @@ Resources:
       HttpMethod: POST
       RestApiId: !Ref GracyChatApi
       ResourceId: !Ref ChatbotResource
-      AuthorizationType: NONE # Consider API keys, IAM auth, or Cognito for production
+      AuthorizationType: NONE
       Integration:
         Type: AWS_PROXY
         IntegrationHttpMethod: POST
